SWDEV-507817 - fix the return type of one of the atomicMin variants


[ROCm/clr commit: e6fb89190a0e7b7d082259753a2ba83fbb0bf924]

+++ hipamd/include/hip/amd_detail/amd_hip_atomic.h
@@ -532,7 +532,7 @@ unsigned int atomicMin_system(unsigned int* address, unsigned int val) {
 
 __device__
 inline
-unsigned long long atomicMin(unsigned long* address, unsigned long val) {
+unsigned long atomicMin(unsigned long* address, unsigned long val) {
 #if defined(__gfx941__)
   return hip_cas_extrema_expander<__ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT>(
     address,
